De-duplicate code by refactoring subprovider classes
